database/requirements.txt
python-decouple
mysql-connector-python
faker